refactor: 렌즈 룰로 통합
