--- src/main/kotlin/core/omock/rule/IsClearFourToFourRule.kt
@@ -1,56 +1,0 @@
-package core.omock.rule
-
-import core.omock.Direction
-import core.omock.Result
-import core.omock.serach.OMockSearch
-
-class IsClearFourToFourRule : OMockRule {
-    override fun validPosition(
-        stoneStates: List<List<Int>>,
-        row: Int,
-        column: Int,
-    ): Boolean {
-        return isClearFourToFour(stoneStates, row, column)
-    }
-
-    private fun isClearFourToFour(
-        stoneStates: List<List<Int>>,
-        row: Int,
-        column: Int,
-    ): Boolean {
-        val visited = OMockSearch.loadMap(stoneStates, row, column)
-        var isClearFourToFourCount = INIT_COUNT
-
-        visited.entries.forEach { (key, result) ->
-            val isReverseResultFirstClear: Boolean = visited[Direction.reverse(key)]?.isFirstClear ?: false
-            val reverseResultCount: Int = visited[Direction.reverse(key)]?.count ?: MIN_REVERSE_COUNT
-
-            if (getClearFourToFourCalculateType(isReverseResultFirstClear, reverseResultCount, result)) {
-                if (++isClearFourToFourCount >= MIN_IS_CLEAR_FOUR_TO_FOUR_COUNT) return true
-            }
-        }
-        return false
-    }
-
-    private fun getClearFourToFourCalculateType(
-        isReverseResultFirstClear: Boolean,
-        reverseResultCount: Int,
-        result: Result,
-    ): Boolean {
-        if (!result.isLastClear) return false
-        if (!result.isFirstClear) return false
-        if (isReverseResultFirstClear && result.count == EDGE_THREE_TO_THREE_COUNT) {
-            return true
-        } else {
-            if (result.count + reverseResultCount == EDGE_THREE_TO_THREE_COUNT) return true
-        }
-        return false
-    }
-
-    companion object {
-        const val INIT_COUNT = 0
-        const val MIN_IS_CLEAR_FOUR_TO_FOUR_COUNT = 4
-        const val MIN_REVERSE_COUNT = 0
-        const val EDGE_THREE_TO_THREE_COUNT = 2
-    }
-}
